% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.R
\name{cancel_worker}
\alias{cancel_worker}
\title{Cancel a running worker}
\usage{
cancel_worker(app, name)
}
\arguments{
\item{app}{An \code{RtuiApp} object.}

\item{name}{Worker name to cancel.}
}
\value{
Invisible \code{app}.
}
\description{
Cancel a running worker
}
