% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{check_terminal}
\alias{check_terminal}
\title{Check if the current terminal supports TUI rendering}
\usage{
check_terminal(call = rlang::caller_env())
}
\value{
\code{TRUE} invisibly if supported; aborts otherwise.
}
\description{
Check if the current terminal supports TUI rendering
}
\keyword{internal}
