% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.R
\name{clear_timer}
\alias{clear_timer}
\title{Cancel a running timer}
\usage{
clear_timer(app, name)
}
\arguments{
\item{app}{An \code{RtuiApp} object.}

\item{name}{Timer name to cancel.}
}
\value{
Invisible \code{app}.
}
\description{
Cancel a running timer
}
