% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-nav.R
\name{collapsible}
\alias{collapsible}
\title{Create a collapsible section}
\usage{
collapsible(title, ..., collapsed = TRUE, id = NULL, classes = NULL)
}
\arguments{
\item{title}{Section title.}

\item{...}{Child widget specs.}

\item{collapsed}{Initial collapsed state.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a collapsible section
}
