% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{collect_form}
\alias{collect_form}
\title{Collect form values from the running app}
\usage{
collect_form(app, field_ids)
}
\arguments{
\item{app}{An \code{RtuiApp} object.}

\item{field_ids}{Character vector of widget ids to query.}
}
\value{
A named list of current values.
}
\description{
Queries all widget ids from a form and returns a named list of their
current values. This is called automatically when the form submit button
is clicked, but you can also call it manually from any handler.
}
