% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-containers.R
\name{container}
\alias{container}
\title{Create a plain block container}
\usage{
container(..., id = NULL, classes = NULL)
}
\arguments{
\item{...}{Child widget specs.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a plain block container
}
