% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{Copy text to system clipboard}
\usage{
copy_to_clipboard(app, text)
}
\arguments{
\item{app}{An \code{RtuiApp} object.}

\item{text}{Text to copy.}
}
\value{
Invisible \code{app}.
}
\description{
Copies the given text to the system clipboard.
}
