% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{dark_toggle}
\alias{dark_toggle}
\title{Toggle dark mode}
\usage{
dark_toggle(app, dark)
}
\arguments{
\item{app}{An \code{RtuiApp} object.}

\item{dark}{Logical. If missing, toggles current mode.}
}
\value{
Invisible \code{app}.
}
\description{
Switches the app between dark and light mode at runtime.
}
