% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-display.R
\name{digits}
\alias{digits}
\title{Create a large digits display widget}
\usage{
digits(value = "", id = NULL, classes = NULL)
}
\arguments{
\item{value}{Text to display in large digits (numbers/colon/space).}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a large digits display widget
}
