% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{event_change}
\alias{event_change}
\title{Wrap a function as a change-event handler}
\usage{
event_change(handler)
}
\arguments{
\item{handler}{A function taking \verb{(event, state)}.}
}
\value{
The handler with class \code{"rtui_change_handler"}.
}
\description{
Wrap a function as a change-event handler
}
