% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{event_click}
\alias{event_click}
\title{Wrap a function as a click-event handler}
\usage{
event_click(handler)
}
\arguments{
\item{handler}{A function taking \verb{(event, state)}.}
}
\value{
The handler with class \code{"rtui_click_handler"}.
}
\description{
Wrap a function as a click-event handler
}
