% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-nav.R
\name{footer}
\alias{footer}
\title{Create a footer widget}
\usage{
footer(id = NULL, classes = NULL)
}
\arguments{
\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a footer widget
}
