% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-nav.R
\name{header}
\alias{header}
\title{Create a header widget}
\usage{
header(show_clock = FALSE, id = NULL, classes = NULL)
}
\arguments{
\item{show_clock}{Show a clock in the header.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a header widget
}
