% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-input.R
\name{list_view}
\alias{list_view}
\title{Create a list view widget}
\usage{
list_view(items, id = NULL, classes = NULL)
}
\arguments{
\item{items}{Character vector of list items.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a list view widget
}
