% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-display.R
\name{loading}
\alias{loading}
\title{Create a loading indicator widget}
\usage{
loading(id = NULL, classes = NULL)
}
\arguments{
\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a loading indicator widget
}
