% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{notify}
\alias{notify}
\title{Show a notification}
\usage{
notify(app, message, severity = c("info", "warning", "error"))
}
\arguments{
\item{app}{An \code{RtuiApp} object.}

\item{message}{Notification message.}

\item{severity}{One of "info", "warning", "error".}
}
\value{
Invisible \code{app}.
}
\description{
Show a notification
}
