% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-display.R
\name{placeholder}
\alias{placeholder}
\title{Create a placeholder widget}
\usage{
placeholder(label = "Placeholder", id = NULL, classes = NULL)
}
\arguments{
\item{label}{Placeholder label text.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a placeholder widget
}
