% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-display.R
\name{pretty_table}
\alias{pretty_table}
\title{Create a pretty table widget (rich-formatted)}
\usage{
pretty_table(df, title = NULL, id = NULL, classes = NULL)
}
\arguments{
\item{df}{A data.frame to display.}

\item{title}{Optional table title.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a pretty table widget (rich-formatted)
}
