% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-input.R
\name{radio_set}
\alias{radio_set}
\title{Create a radio set (group of radio buttons)}
\usage{
radio_set(..., id = NULL, classes = NULL)
}
\arguments{
\item{...}{Radio button specs.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a radio set (group of radio buttons)
}
