% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{RtuiState}
\alias{RtuiState}
\title{Mutable TUI state object}
\description{
Mutable TUI state object

Mutable TUI state object
}
\details{
An R6 class for managing mutable app state. Created by \code{\link[=tui_state]{tui_state()}}.
Supports reactive bindings that auto-update widgets when values change.
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{Access the raw state data list.}

\item{\code{app}}{Access the running app (for calling \code{update()}, etc.)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RtuiState-new}{\code{RtuiState$new()}}
\item \href{#method-RtuiState-get}{\code{RtuiState$get()}}
\item \href{#method-RtuiState-set}{\code{RtuiState$set()}}
\item \href{#method-RtuiState-as_list}{\code{RtuiState$as_list()}}
\item \href{#method-RtuiState-print}{\code{RtuiState$print()}}
\item \href{#method-RtuiState-clone}{\code{RtuiState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RtuiState-new"></a>}}
\if{latex}{\out{\hypertarget{method-RtuiState-new}{}}}
\subsection{Method \code{new()}}{
Create a new state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RtuiState$new(initial = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{A list of initial state values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RtuiState-get"></a>}}
\if{latex}{\out{\hypertarget{method-RtuiState-get}{}}}
\subsection{Method \code{get()}}{
Get a value by key.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RtuiState$get(key, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Character string key.}

\item{\code{default}}{Value to return if key is not found.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RtuiState-set"></a>}}
\if{latex}{\out{\hypertarget{method-RtuiState-set}{}}}
\subsection{Method \code{set()}}{
Set a value by key. Fires reactive bindings if the value changed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RtuiState$set(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Character string key.}

\item{\code{value}}{The value to store.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RtuiState-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-RtuiState-as_list}{}}}
\subsection{Method \code{as_list()}}{
Return state as a plain list (excluding internal keys).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RtuiState$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RtuiState-print"></a>}}
\if{latex}{\out{\hypertarget{method-RtuiState-print}{}}}
\subsection{Method \code{print()}}{
Print the state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RtuiState$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RtuiState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RtuiState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RtuiState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
