% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-input.R
\name{selection_list}
\alias{selection_list}
\title{Create a selection list (multi-select)}
\usage{
selection_list(items, id = NULL, classes = NULL)
}
\arguments{
\item{items}{Character vector of items.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Like \code{\link[=option_list]{option_list()}} but allows the user to toggle multiple items on/off.
}
\examples{
\dontrun{
quick_app(
  layout = vstack(
    static("Select permissions:"),
    selection_list(items = c("Read", "Write", "Execute"),
                   id = "perms"),
    static("", id = "chosen"),
    id = "root"
  ),
  on_change = list(
    perms = function(event, state) {
      update(state$app, "chosen",
             content = paste("Selected:", toString(event$value)))
      state
    }
  )
)
}

}
