% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-input.R
\name{switch_input}
\alias{switch_input}
\title{Create a switch (toggle) widget}
\usage{
switch_input(value = FALSE, id = NULL, classes = NULL, tooltip = NULL)
}
\arguments{
\item{value}{Initial state (logical).}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}

\item{tooltip}{Optional tooltip text shown on hover.}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a switch (toggle) widget
}
