% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-nav.R
\name{tabs}
\alias{tabs}
\title{Create a tabbed pane container}
\usage{
tabs(..., id = NULL, classes = NULL)
}
\arguments{
\item{...}{Tab pane specs created with \code{tab_pane()}.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a tabbed pane container
}
