% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{tui_state}
\alias{tui_state}
\title{Create a mutable TUI state object}
\usage{
tui_state(initial = list())
}
\arguments{
\item{initial}{A list of initial state values.}
}
\value{
An R6 object of class \code{RtuiState}.
}
\description{
Create a mutable TUI state object
}
