% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{tui_vbox}
\alias{tui_vbox}
\title{Vertical box layout}
\usage{
tui_vbox(...)
}
\arguments{
\item{...}{Child components built with layout or component functions.}
}
\value{
A \code{rtui_component} list node of type \code{"vbox"}.
}
\description{
Stacks child components vertically.
}
