% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tuiIsolate}
\alias{tuiIsolate}
\title{Evaluate an expression in isolation}
\usage{
tuiIsolate(expr)
}
\arguments{
\item{expr}{An expression to evaluate.}
}
\value{
The evaluated value of \code{expr}.
}
\description{
Evaluates \code{expr} without publishing reactive read-change information for the
current server cycle. This is useful when you want the current value of a
reactive object without making it act as an event trigger.
}
