% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tuiObserve}
\alias{tuiObserve}
\title{Observe reactive values}
\usage{
tuiObserve(expr)
}
\arguments{
\item{expr}{An expression to evaluate.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Evaluates \code{expr} every server run.
}
