% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tuiOutputNumeric}
\alias{tuiOutputNumeric}
\title{Numeric output component}
\usage{
tuiOutputNumeric(outputId)
}
\arguments{
\item{outputId}{A single character string naming the output id.}
}
\value{
A \code{rtuiComponent} list node of type \code{"outputNumeric"}.
}
\description{
Displays the value of \verb{output$<outputId>} as numeric text.
}
