% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tuiReactive}
\alias{tuiReactive}
\title{Create a reactive expression}
\usage{
tuiReactive(expr)
}
\arguments{
\item{expr}{An expression producing a reactive value.}
}
\value{
A reactive object (function) to be called with \verb{()}.
}
\description{
Creates a reactive object that re-evaluates \code{expr} when called during server
execution.
}
