% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tuiReactiveEvent}
\alias{tuiReactiveEvent}
\title{Create an event-driven reactive expression}
\usage{
tuiReactiveEvent(event, expr, runAtInit = FALSE)
}
\arguments{
\item{event}{An \verb{input$<id>} / \code{input[["id"]]} reference, or a reactive
call/object from \code{\link[=tuiReactive]{tuiReactive()}}, \code{\link[=tuiReactiveVal]{tuiReactiveVal()}}, or
\code{\link[=tuiReactiveEvent]{tuiReactiveEvent()}}.}

\item{expr}{An expression evaluated when \code{event} triggers.}

\item{runAtInit}{A single logical value. If \code{TRUE}, evaluate once during app
initialization.}
}
\value{
A reactive object (function) to be called with \verb{()}.
}
\description{
Creates a reactive object that updates only when \code{event} is triggered.
}
