% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.R
\name{tuiReactiveTimer}
\alias{tuiReactiveTimer}
\title{Create a timer-driven reactive value}
\usage{
tuiReactiveTimer(intervalMs = 1000L)
}
\arguments{
\item{intervalMs}{Timer interval in milliseconds (must be > 0).}
}
\value{
A reactive object (function) to be called with \verb{()}.
}
\description{
Creates a reactive object that increments at a fixed interval.
}
