% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tuiReactiveVal}
\alias{tuiReactiveVal}
\title{Create a mutable reactive value}
\usage{
tuiReactiveVal(value = NULL)
}
\arguments{
\item{value}{Initial value.}
}
\value{
A reactive value function.
}
\description{
Creates a reactive value container. Call without arguments to read, and with
one argument to update.
}
