% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tuiRenderNumeric}
\alias{tuiRenderNumeric}
\title{Create a numeric renderer for \code{output$...}}
\usage{
tuiRenderNumeric(expr, digits = NULL)
}
\arguments{
\item{expr}{An expression returning numeric-compatible output.}

\item{digits}{Optional number of digits after the decimal point.}
}
\value{
A numeric renderer object for assignment in \code{output$...}.
}
\description{
Create a numeric renderer for \code{output$...}
}
