% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-inputs.R
\name{tuiUpdateCheckboxInput}
\alias{tuiUpdateCheckboxInput}
\title{Update a checkbox input value from server logic}
\usage{
tuiUpdateCheckboxInput(id, value)
}
\arguments{
\item{id}{Input id passed to \code{\link[=tuiInputCheckbox]{tuiInputCheckbox()}}.}

\item{value}{New logical value.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Sets the current value of an existing \code{\link[=tuiInputCheckbox]{tuiInputCheckbox()}} input.
}
