% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-inputs.R
\name{tuiUpdateTextInput}
\alias{tuiUpdateTextInput}
\title{Update a text input value from server logic}
\usage{
tuiUpdateTextInput(id, value)
}
\arguments{
\item{id}{Input id passed to \code{\link[=tuiInputText]{tuiInputText()}}.}

\item{value}{New text value.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Sets the current value of an existing \code{\link[=tuiInputText]{tuiInputText()}} input.
}
