% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update a widget by id}
\usage{
update(app, id, ...)
}
\arguments{
\item{app}{An \code{RtuiApp} object.}

\item{id}{Widget id to update.}

\item{...}{Named properties to set on the widget.}
}
\value{
Invisible \code{app}.
}
\description{
Update a widget by id
}
