% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-input.R
\name{option_list}
\alias{option_list}
\title{Create an option list widget}
\usage{
option_list(items, id = NULL, classes = NULL)
}
\arguments{
\item{items}{Character vector of option items.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create an option list widget
}
