% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-display.R
\name{progress_bar}
\alias{progress_bar}
\title{Create a progress bar widget}
\usage{
progress_bar(
  total = 100,
  progress = 0,
  show_eta = TRUE,
  show_percentage = TRUE,
  id = NULL,
  classes = NULL
)
}
\arguments{
\item{total}{Total value (numeric).}

\item{progress}{Current progress value (numeric).}

\item{show_eta}{Show estimated time of arrival.}

\item{show_percentage}{Show percentage.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a progress bar widget
}
