% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-input.R
\name{selection_list}
\alias{selection_list}
\title{Create a selection list (multi-select)}
\usage{
selection_list(items, id = NULL, classes = NULL)
}
\arguments{
\item{items}{Character vector of items.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a selection list (multi-select)
}
