% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets-display.R
\name{sparkline}
\alias{sparkline}
\title{Create a sparkline widget}
\usage{
sparkline(data, id = NULL, classes = NULL)
}
\arguments{
\item{data}{Numeric vector of values.}

\item{id}{Optional widget id.}

\item{classes}{Optional CSS classes (character vector).}
}
\value{
An \code{rtui_spec} list.
}
\description{
Create a sparkline widget
}
