% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{tuiColumn}
\alias{tuiColumn}
\title{Vertical box layout}
\usage{
tuiColumn(...)
}
\arguments{
\item{...}{Child components built with layout or component functions.}
}
\value{
A \code{rtuiComponent} list node of type \code{"column"}.
}
\description{
Stacks child components vertically.
}
