% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tuiInputButton}
\alias{tuiInputButton}
\title{Button component}
\usage{
tuiInputButton(label, id)
}
\arguments{
\item{label}{A character string shown as the button label.}

\item{id}{A character string used as the input key (\verb{input$<id>}).}
}
\value{
A \code{rtuiComponent} list node of type \code{"button"}.
}
\description{
A focusable button. When activated (Enter key), it increments \code{input$id}
and triggers a server update.
}
