% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tuiInputCheckbox}
\alias{tuiInputCheckbox}
\title{Checkbox input component}
\usage{
tuiInputCheckbox(label, id, value = FALSE)
}
\arguments{
\item{label}{A character string shown as the checkbox label.}

\item{id}{A character string used as the input key (\verb{input$<id>}).}

\item{value}{A single logical value used as the initial/default checkbox value.}
}
\value{
A \code{rtuiComponent} list node of type \code{"checkbox"}.
}
\description{
A focusable checkbox. Toggling it updates \verb{input$<id>} and triggers
a server update.
}
