% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive-api.R
\name{tuiObserve}
\alias{tuiObserve}
\title{Observe reactive values}
\usage{
tuiObserve(expr)
}
\arguments{
\item{expr}{An expression to evaluate.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Registers an observer that re-evaluates \code{expr} when dependencies invalidate.
}
