% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{tuiObserveEvent}
\alias{tuiObserveEvent}
\title{Observe selected input/reactive events}
\usage{
tuiObserveEvent(event, expr, runAtInit = FALSE)
}
\arguments{
\item{event}{An \verb{input$<id>} / \code{input[["id"]]} reference, or a reactive
call/object from \code{\link[=tuiReactive]{tuiReactive()}}, \code{\link[=tuiReactiveVal]{tuiReactiveVal()}}, or
\code{\link[=tuiReactiveEvent]{tuiReactiveEvent()}}.}

\item{expr}{An expression evaluated when \code{event} triggers.}

\item{runAtInit}{A single logical value. If \code{TRUE}, evaluate once during app
initialization.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Registers an observer that evaluates \code{expr} only when \code{event} is triggered.
}
