% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tuiOutputText}
\alias{tuiOutputText}
\title{Text output component}
\usage{
tuiOutputText(outputId)
}
\arguments{
\item{outputId}{A single character string naming the output id.}
}
\value{
A \code{rtuiComponent} list node of type \code{"outputText"}.
}
\description{
Displays the value of \verb{output$<outputId>} as text.
}
