% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tuiRenderText}
\alias{tuiRenderText}
\title{Text output component}
\usage{
tuiRenderText(key)
}
\arguments{
\item{key}{An \verb{output$<key>} or \code{output[["key"]]} reference.}
}
\value{
A \code{rtuiComponent} list node of type \code{"text"}.
}
\description{
Displays the value of an \code{output} key as text. The display updates
automatically whenever the server updates that output key.
}
