% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tuiRenderText}
\alias{tuiRenderText}
\title{Create a text renderer for \code{output$...}}
\usage{
tuiRenderText(expr)
}
\arguments{
\item{expr}{An expression returning text-compatible output.}
}
\value{
A text renderer object for assignment in \code{output$...}.
}
\description{
Create a text renderer for \code{output$...}
}
